Fix putative problem in tests and update patch
NPM audit and dependabot report a critical vulnerability in the tests
because they were using `bats-file`.  A version of this dependency was
apparently hacked and the hacked version is malware.  I'm quite sure
we are _not_ using the hacked version but this does not seem to matter
to the robots-that-be.  To stop getting scary messages, I have dropped
the dependency (we didn't really use it for much and the affected
tests were easily redesigned).
Also committed the nimbella-cli patch that should have been committed
last time.